requirements.txt
Babel==2.9.1
certifi==2021.10.8
charset-normalizer==2.0.11
idna==3.3
Pillow==9.0.1
pytz==2021.3
requests==2.27.1
tkcalendar==1.6.1
urllib3==1.26.8
